+++ db/migrate/20160214104259_add_scope_exclude_file_extensions_to_profiles.rb
@@ -1,0 +1,5 @@
+class AddScopeExcludeFileExtensionsToProfiles < ActiveRecord::Migration
+  def change
+    add_column :profiles, :scope_exclude_file_extensions, :text
+  end
+end