+++ db/migrate/20160303170652_add_http_authentication_type_to_profiles.rb
@@ -1,0 +1,5 @@
+class AddHttpAuthenticationTypeToProfiles < ActiveRecord::Migration
+    def change
+        add_column :profiles, :http_authentication_type, :string
+    end
+end